+++ server/webapp/src/main/java/cloud/qasino/games/database/repository/CardMoveRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CardMoveRepository extends JpaRepository<CardMove, Integer> {
+public interface CardMoveRepository extends JpaRepository<CardMove, Long> {
 
 
     List<CardMove> findByTurn(Turn turn);
